+++ lib/postgrex/type_server.ex
@@ -10,9 +10,9 @@ defmodule Postgrex.TypeServer do
   @doc """
   Starts a type server.
   """
-  @spec start_link(module, pid) :: GenServer.on_start
-  def start_link(module, starter) do
-    GenServer.start_link(__MODULE__, {module, starter})
+  @spec start_link(module, pid, keyword) :: GenServer.on_start
+  def start_link(module, starter, opts) do
+    GenServer.start_link(__MODULE__, {module, starter}, opts)
   end
 
   @doc """