+++ test/transaction_test.exs
@@ -10,8 +10,16 @@ defmodule TransactionTest do
         :transaction -> :strict
         :savepoint   -> :naive
       end
-    opts = [ database: "postgrex_test", transactions: transactions, idle: :active,
-             backoff_type: :stop, prepare: context[:prepare] || :named]
+
+    opts = [
+      database: "postgrex_test",
+      transactions: transactions,
+      idle: :active,
+      backoff_type: :stop,
+      prepare: context[:prepare] || :named,
+      max_restarts: 0
+    ]
+
     {:ok, pid} = P.start_link(opts)
     {:ok, [pid: pid]}
   end
@@ -32,11 +40,14 @@ defmodule TransactionTest do
   test "connection works after failure during rollback transaction", context do
     assert transaction(fn(conn) ->
       assert {:error, %Postgrex.Error{postgres: %{code: :unique_violation}}} =
-       P.query(conn, "insert into uniques values (1), (1);", [])
-     assert {:error, %Postgrex.Error{postgres: %{code: :in_failed_sql_transaction}}} =
-       P.query(conn, "SELECT 42", [])
-       P.rollback(conn, :oops)
+               P.query(conn, "insert into uniques values (1), (1);", [])
+
+      assert {:error, %Postgrex.Error{postgres: %{code: :in_failed_sql_transaction}}} =
+               P.query(conn, "SELECT 42", [])
+
+      P.rollback(conn, :oops)
     end) == {:error, :oops}
+
     assert [[42]] = query("SELECT 42", [])
   end
 
@@ -44,12 +55,11 @@ defmodule TransactionTest do
   test "query begin returns error", context do
     Process.flag(:trap_exit, true)
 
-    capture_log fn ->
-      assert (%Postgrex.Error{message: "unexpected postgres status: transaction"} = err) = query("BEGIN", [])
-
+    assert capture_log(fn ->
+      assert %Postgrex.Error{message: "unexpected postgres status: transaction"} = query("BEGIN", [])
       pid = context[:pid]
-      assert_receive {:EXIT, ^pid, {:shutdown, ^err}}
-    end
+      assert_receive {:EXIT, ^pid, :killed}
+    end) =~ "** (Postgrex.Error) unexpected postgres status: transaction"
   end
 
   @tag mode: :transaction
@@ -57,51 +67,16 @@ defmodule TransactionTest do
     Process.flag(:trap_exit, true)
 
     assert transaction(fn(conn) ->
-      capture_log fn ->
-        assert {:error, %Postgrex.Error{message: "unexpected postgres status: idle"} = err} =
+      assert capture_log(fn ->
+        assert {:error, %Postgrex.Error{message: "unexpected postgres status: idle"}} =
           P.query(conn, "ROLLBACK", [])
 
         pid = context[:pid]
-        assert_receive {:EXIT, ^pid, {:shutdown, ^err}}
-      end
+        assert_receive {:EXIT, ^pid, :killed}
+      end) =~ "** (Postgrex.Error) unexpected postgres status: idle"
+
       :hi
     end) == {:error, :rollback}
-  end
-
-  @tag mode: :transaction
-  test "checkout when in transaction disconnects", context do
-    Process.flag(:trap_exit, true)
-
-    pid = context[:pid]
-    :sys.replace_state(pid,
-      fn(%{mod_state: %{state: state} = mod} = conn) ->
-        %{conn | mod_state: %{mod | state: %{state | postgres: :transaction}}}
-      end)
-    capture_log fn ->
-      assert {{:shutdown,
-          %Postgrex.Error{message: "unexpected postgres status: transaction"} = err}, _} =
-        catch_exit(query("SELECT 42", []))
-
-      assert_receive {:EXIT, ^pid, {:shutdown, ^err}}
-    end
-  end
-
-  @tag mode: :transaction
-  test "ping when transaction state mismatch disconnects" do
-    Process.flag(:trap_exit, true)
-
-    opts = [ database: "postgrex_test", transactions: :strict,
-             idle_timeout: 10, backoff_type: :stop ]
-    {:ok, pid} = P.start_link(opts)
-
-    capture_log fn ->
-      :sys.replace_state(pid,
-        fn(%{mod_state: %{state: state} = mod} = conn) ->
-          %{conn | mod_state: %{mod | state: %{state | postgres: :transaction}}}
-        end)
-      assert_receive {:EXIT, ^pid, {:shutdown,
-          %Postgrex.Error{message: "unexpected postgres status: transaction"}}}
-    end
   end
 
   @tag mode: :transaction
@@ -286,7 +261,9 @@ defmodule TransactionTest do
 
   @tag mode: :transaction
   test "savepoint query disconnects on release savepoint error", context do
-    assert transaction(fn(conn) ->
+    Process.flag(:trap_exit, true)
+
+    assert transaction(fn conn ->
       assert {:error, %Postgrex.Error{postgres: %{code: :invalid_savepoint_specification}}} =
         P.query(conn, "RELEASE SAVEPOINT postgrex_query", [], [mode: :savepoint])
 
@@ -464,16 +441,21 @@ defmodule TransactionTest do
     assert DBConnection.status(pid, opts) == :idle
     assert query("SELECT 42", []) == [[42]]
     assert DBConnection.status(pid, opts) == :idle
-    {conn, _} = DBConnection.begin!(pid, opts)
-    assert DBConnection.status(conn, opts) == :transaction
-    assert {:error, %Postgrex.Error{postgres: %{code: :unique_violation}}} =
-        P.query(conn, "insert into uniques values (1), (1);", [], opts)
-    assert DBConnection.status(conn, opts) == :error
-    assert {:error, %Postgrex.Error{postgres: %{code: :in_failed_sql_transaction}}} =
-        P.query(conn, "SELECT 42", [], opts)
-    assert DBConnection.status(conn, opts) == :error
-    assert {:error, %DBConnection.TransactionError{status: :error}} =
-      DBConnection.commit(conn, opts)
+
+    assert DBConnection.transaction(pid, fn conn ->
+             assert DBConnection.status(conn, opts) == :transaction
+
+             assert {:error, %Postgrex.Error{postgres: %{code: :unique_violation}}} =
+                      P.query(conn, "insert into uniques values (1), (1);", [], opts)
+
+             assert DBConnection.status(conn, opts) == :error
+
+             assert {:error, %Postgrex.Error{postgres: %{code: :in_failed_sql_transaction}}} =
+                      P.query(conn, "SELECT 42", [], opts)
+
+             assert DBConnection.status(conn, opts) == :error
+           end, opts) == {:error, :rollback}
+
     assert DBConnection.status(pid, opts) == :idle
     assert query("SELECT 42", []) == [[42]]
     assert DBConnection.status(pid) == :idle